Move queries to new file
